pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "scrare-refine"
version = "0.1.0"
description = "Rare-aware refinement for single-cell annotation via scANVI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "anndata>=0.9",
    "matplotlib>=3.7",
    "numpy>=1.23",
    "pandas>=1.5",
    "psutil>=5.9",
    "PyYAML>=6.0",
    "scanpy>=1.9",
    "scikit-learn>=1.2",
    "scipy>=1.10",
    "scvi-tools>=0.20",
    "torch>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["scrare_refine*"]
